Refactor Permissions-Policy header to use PermissionsPolicyDirective enum
