+++ app/src/Http/SecurityHeaders/PermissionsPolicy/PermissionsPolicyOrigin.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace MichalSpacekCz\Http\SecurityHeaders;
+namespace MichalSpacekCz\Http\SecurityHeaders\PermissionsPolicy;
 
 enum PermissionsPolicyOrigin: string
 {
